move flag tests
